Refactor Phase 2: move leaf modules into target subdirs
Moves each leaf module into the subdirectory that matches its role
in the customer flow:

  mapper/semantic_matching/property_matcher.py  (was semantic_matcher.py)
  mapper/v1_to_v2/bigquery_transformer.py
  mapper/v1_to_v2/debezium_translator.py
  mapper/v1_to_v2/http_transformer.py
  migration/offset_manager.py
  output/summary_report.py                       (was summary.py)
  output/terraform_generator.py

Renames the internal class SemanticMatcher -> SemanticPropertyMatcher.
The user-facing custom-matcher extension point (loads a class named
`SemanticMatcher` from a user-provided file) is intentionally unchanged
in connector_comparator.py so existing customer extensions keep working.

All intra-package imports updated. No logic changes; --help output and
output JSON schema are unchanged. Also adds .gitignore for __pycache__
and build artifacts.
